refactor: extract STT providers into WhisperDesk.Stt module
- Created src/stt/WhisperDesk.Stt.Contract (interfaces + data models)
- Created src/stt/WhisperDesk.Stt (Azure, Volcengine implementations)
- Core now depends on Contract only, not on STT implementations
- WhisperDesk.Stt has zero dependency on Core
- DI composition done in UI layer (Composition Root pattern)
