+++ src/stt/WhisperDesk.Stt.Contract/WhisperDesk.Stt.Contract.csproj
@@ -1,0 +1,8 @@
+<Project Sdk="Microsoft.NET.Sdk">
+  <PropertyGroup>
+    <TargetFramework>net9.0</TargetFramework>
+    <Nullable>enable</Nullable>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <RootNamespace>WhisperDesk.Stt.Contract</RootNamespace>
+  </PropertyGroup>
+</Project>
