+++ src/stt/WhisperDesk.Stt/Volcengine/VolcengineSttConfig.cs
@@ -1,0 +1,7 @@
+namespace WhisperDesk.Stt.Volcengine;
+
+public class VolcengineSttConfig
+{
+    public string ApiKey { get; set; } = string.Empty;
+    public string ResourceId { get; set; } = "volc.seedasr.sauc.duration";
+}
